Renamed to uniform point spacing
