+++ noether_gui/CMakeLists.txt
@@ -60,7 +60,7 @@ add_library(${PROJECT_NAME} SHARED
   src/widgets/tool_path_modifiers/uniform_orientation_modifier_widget.cpp
   src/widgets/tool_path_modifiers/offset_modifier_widget.cpp
   src/widgets/tool_path_modifiers/spline_modifier_widget.cpp
-  src/widgets/tool_path_modifiers/point_spacing_modifier_widget.cpp
+  src/widgets/tool_path_modifiers/uniform_point_spacing_modifier_widget.cpp
   # Mesh Modifiers
   src/widgets/mesh_modifiers/plane_projection_modifier_widget.cpp
   src/widgets/mesh_modifiers/euclidean_clustering_modifier_widget.cpp
