+++ noether_tpp/CMakeLists.txt
@@ -51,7 +51,7 @@ add_library(${PROJECT_NAME} SHARED
   src/tool_path_modifiers/uniform_orientation_modifier.cpp
   src/tool_path_modifiers/offset_modifier.cpp
   src/tool_path_modifiers/spline_modifier.cpp
-  src/tool_path_modifiers/point_spacing_modifier.cpp
+  src/tool_path_modifiers/uniform_point_spacing_modifier.cpp
   # Tool Path Planners
   src/tool_path_planners/multi_tool_path_planner.cpp
   src/tool_path_planners/edge/edge_planner.cpp
